Externalisation of signal busLines and of the function fetchBusLines: fetch bus lines after delete.
